Tue, Dec 6, 2022, 1:17 AM +00:00

+++ misc/avg.c
@@ -4,6 +4,6 @@
 int main(void)
 {
     float x = (27 + 28 + 28)/3;
-    int y = ceil (x);
-    printf("%i\n",y);
+    float y = ceil (x);
+    printf("%f\n",y);
 }
